dashboard -> explore correct global css and setCredentials on student login
