+++ packages/client/styles/global.css
@@ -23,7 +23,8 @@ body {
 
 * {
   box-sizing: border-box;
-  font-family: "Poppins";
+  font-family: "Poppins", sans-serif;
+  transition: all 0.2s ease-in-out;
 }
 
 a {
